docs from EC2-console sess to debug connectivity
Documentation - I was never able to fully resolve the Refuse Connection error.  I used that as the subject starter for my Chat with ai.  I figured the first step to idempotent code is getting to run the first time. . .Chat Gpt5 concluded that my port was not listening, and the fixes I was given had me running a sanity check on the EC2 console and making some changes in the security-group settings to allow inbound traffic, as the previous setting was interfering with listening at port 80.  Following the steps (as shown in the debugging.txt) I was able to get my instance working at my public URL; but I was not able to get as far as running the app there.  Whatever the Cloud equivalent is of the Rabbit Hole, that's where I was.